pyproject.toml
[project]
name = "livecold-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pathway[all]",
    "python-dotenv",
    "paho-mqtt",
    "sentence-transformers",
    "litellm",
    "unstructured",
    "tiktoken",
]
